% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcts_main.R
\name{mcts}
\alias{mcts}
\title{Simulate Multiple Choice Test}
\usage{
mcts(x, ...)
}
\arguments{
\item{x}{the inputted test. This can be of types: \cr
matrix - writing test directly in R \cr
character - reading from file name \cr
data.frame - if file has already been read}

\item{...}{used to pass repetitions and header for file input, as well as truncation boolean.}
}
\value{
A list with class 'test' displaying the format of the test as well as the simulated results in both a raw and tabulated form.
}
\description{
The main function to be used in simulating multiple choice tests. \cr

Formatting for input x is imperative. For a test of n questions, mcts expects an n x 3 matrix with each row representing one question. \cr
The first column determines the question type; "normal" - one correct answer out of p questions; "neg" - one correct answer out of p possible answers where you lose 1 mark for an incorrect answer;
"multi" - q correct answers out of p possible answers with 1 mark gained for each correct answer. The second column determines the number of possible answers p.
The third column is reserved for the special case "multi" and determines the number of correct answers q.\cr \cr
This test can be inputted either by matrix, file (character) or data.frame.
}
