% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcts_outs.R
\name{score}
\alias{score}
\title{Calculate probability of scoring over v percent on a test x}
\usage{
score(x, v = 0.4)
}
\arguments{
\item{x}{the input test}

\item{v}{the threshold score percentage (defaults to 0.40)}
}
\value{
The proportion of students who would score at least v percent by random guessing.
}
\description{
Calculate probability of scoring over v percent on a test x
}
